@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.7+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -26,28 +26,34 @@ dependencies:
   flutter_localizations:
     sdk: flutter
   auto_size_text: 3.0.0
-  cached_network_image: 3.1.0+1
-  firebase_auth: 3.2.0
-  firebase_core: 1.12.0
-  firebase_core_platform_interface: 4.4.1
-  firebase_core_web: 1.5.4
+  cached_network_image: 3.2.1
+  firebase_auth: 3.4.1
+  firebase_auth_platform_interface: 6.3.1
+  firebase_auth_web: 3.3.19
+  firebase_core: 1.19.1
+  firebase_core_platform_interface: 4.4.3
+  firebase_core_web: 1.6.6
   flutter_spinkit: 5.1.0
   font_awesome_flutter: 10.1.0
-  google_fonts: 2.2.0
-  google_sign_in: 5.2.1
+  go_router: 3.1.0
+  google_fonts: 3.0.1
+  google_sign_in: 5.4.0
+  google_sign_in_android: 6.0.1
+  google_sign_in_ios: 5.4.0
+  google_sign_in_platform_interface: 2.2.0
+  google_sign_in_web: 0.10.2
   intl: 0.17.0
-  json_path: 0.3.1
-  page_transition: 2.0.4
-  photo_view: ^0.13.0
-  rxdart: 0.26.0
-  shared_preferences: 2.0.11
-  sign_in_with_apple:
-    git:
-      url: https://github.com/asmengistu/dart_packages
-      ref: ab18d53a80e0a4bb90f1ad84f37f1b408930f28d
-      path: packages/sign_in_with_apple/sign_in_with_apple/
-  timeago: 3.1.0
-  url_launcher: 6.0.15
+  json_path: 0.4.1
+  page_transition: 2.0.9
+  photo_view: 0.14.0
+  plugin_platform_interface: 2.1.2
+  rxdart: 0.27.4
+  shared_preferences: 2.0.15
+  sign_in_with_apple: 4.1.0
+  sign_in_with_apple_platform_interface: 1.0.0
+  sign_in_with_apple_web: 1.0.1
+  timeago: 3.2.2
+  url_launcher: 6.1.5
   webviewx: 0.2.1
 
 
@@ -56,7 +62,7 @@ dependencies:
   cupertino_icons: ^1.0.0
 
 dev_dependencies:
-  flutter_launcher_icons: 0.9.2
+  flutter_launcher_icons: 0.9.3
 
   flutter_test:
     sdk: flutter
