@@ -9,18 +9,19 @@ export 'anonymous_auth.dart';
 export 'apple_auth.dart';
 export 'email_auth.dart';
 export 'google_auth.dart';
+export 'jwt_token_auth.dart';
 
 /// Tries to sign in or create an account using Firebase Auth.
 /// Returns the User object if sign in was successful.
-Future<User> signInOrCreateAccount(
-    BuildContext context, Future<UserCredential> Function() signInFunc) async {
+Future<User?> signInOrCreateAccount(
+    BuildContext context, Future<UserCredential?> Function() signInFunc) async {
   try {
     final userCredential = await signInFunc();
-    return userCredential.user;
+    return userCredential?.user;
   } on FirebaseAuthException catch (e) {
     ScaffoldMessenger.of(context).hideCurrentSnackBar();
     ScaffoldMessenger.of(context).showSnackBar(
-      SnackBar(content: Text('Error: ${e.message}')),
+      SnackBar(content: Text('Error: ${e.message!}')),
     );
     return null;
   }
@@ -28,7 +29,7 @@ Future<User> signInOrCreateAccount(
 
 Future signOut() {
   _currentJwtToken = '';
-  FirebaseAuth.instance.signOut();
+  return FirebaseAuth.instance.signOut();
 }
 
 Future deleteUser(BuildContext context) async {
@@ -50,13 +51,14 @@ Future deleteUser(BuildContext context) async {
   }
 }
 
-Future resetPassword({String email, BuildContext context}) async {
+Future resetPassword(
+    {required String email, required BuildContext context}) async {
   try {
     await FirebaseAuth.instance.sendPasswordResetEmail(email: email);
   } on FirebaseAuthException catch (e) {
     ScaffoldMessenger.of(context).hideCurrentSnackBar();
     ScaffoldMessenger.of(context).showSnackBar(
-      SnackBar(content: Text('Error: ${e.message}')),
+      SnackBar(content: Text('Error: ${e.message!}')),
     );
     return null;
   }
@@ -68,7 +70,7 @@ Future resetPassword({String email, BuildContext context}) async {
 Future sendEmailVerification() async =>
     currentUser?.user?.sendEmailVerification();
 
-String _currentJwtToken = '';
+String? _currentJwtToken = '';
 
 String get currentUserEmail => currentUser?.user?.email ?? '';
 
@@ -85,14 +87,14 @@ String get currentJwtToken => _currentJwtToken ?? '';
 bool get currentUserEmailVerified => currentUser?.user?.emailVerified ?? false;
 
 // Set when using phone verification (after phone number is provided).
-String _phoneAuthVerificationCode;
+String? _phoneAuthVerificationCode;
 // Set when using phone sign in in web mode (ignored otherwise).
-ConfirmationResult _webPhoneAuthConfirmationResult;
+ConfirmationResult? _webPhoneAuthConfirmationResult;
 
 Future beginPhoneAuth({
-  BuildContext context,
-  String phoneNumber,
-  VoidCallback onCodeSent,
+  required BuildContext context,
+  required String phoneNumber,
+  required VoidCallback onCodeSent,
 }) async {
   if (kIsWeb) {
     _webPhoneAuthConfirmationResult =
@@ -120,7 +122,7 @@ Future beginPhoneAuth({
     },
     verificationFailed: (e) {
       ScaffoldMessenger.of(context).showSnackBar(SnackBar(
-        content: Text('Error: ${e.message}'),
+        content: Text('Error: ${e.message!}'),
       ));
     },
     codeSent: (verificationId, _) {
@@ -132,15 +134,15 @@ Future beginPhoneAuth({
 }
 
 Future verifySmsCode({
-  BuildContext context,
-  String smsCode,
+  required BuildContext context,
+  required String smsCode,
 }) async {
   if (kIsWeb) {
     return signInOrCreateAccount(
-        context, () => _webPhoneAuthConfirmationResult.confirm(smsCode));
+        context, () => _webPhoneAuthConfirmationResult!.confirm(smsCode));
   } else {
     final authCredential = PhoneAuthProvider.credential(
-        verificationId: _phoneAuthVerificationCode, smsCode: smsCode);
+        verificationId: _phoneAuthVerificationCode!, smsCode: smsCode);
     return signInOrCreateAccount(
       context,
       () => FirebaseAuth.instance.signInWithCredential(authCredential),
