@@ -6,7 +6,7 @@ import 'package:flutter_spinkit/flutter_spinkit.dart';
 import 'package:google_fonts/google_fonts.dart';
 
 class CountyWidget extends StatefulWidget {
-  const CountyWidget({Key key}) : super(key: key);
+  const CountyWidget({Key? key}) : super(key: key);
 
   @override
   _CountyWidgetState createState() => _CountyWidgetState();
@@ -16,23 +16,33 @@ class _CountyWidgetState extends State<CountyWidget> {
   final scaffoldKey = GlobalKey<ScaffoldState>();
 
   @override
+  void initState() {
+    super.initState();
+
+    WidgetsBinding.instance.addPostFrameCallback((_) => setState(() {}));
+  }
+
+  @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      key: scaffoldKey,
-      backgroundColor: FlutterFlowTheme.of(context).primaryBackground,
-      body: SafeArea(
-        child: GestureDetector(
-          onTap: () => FocusScope.of(context).unfocus(),
-          child: FlutterFlowWebView(
-            url: 'https://www.gchd.us/coronavirus/',
-            bypass: false,
-            width: MediaQuery.of(context).size.width,
-            height: MediaQuery.of(context).size.height * 1,
-            verticalScroll: false,
-            horizontalScroll: false,
+    return Title(
+        title: 'county',
+        color: FlutterFlowTheme.of(context).primaryColor,
+        child: Scaffold(
+          key: scaffoldKey,
+          backgroundColor: FlutterFlowTheme.of(context).primaryBackground,
+          body: SafeArea(
+            child: GestureDetector(
+              onTap: () => FocusScope.of(context).unfocus(),
+              child: FlutterFlowWebView(
+                url: 'https://www.gchd.us/coronavirus/',
+                bypass: false,
+                width: MediaQuery.of(context).size.width,
+                height: MediaQuery.of(context).size.height * 1,
+                verticalScroll: false,
+                horizontalScroll: false,
+              ),
+            ),
           ),
-        ),
-      ),
-    );
+        ));
   }
 }
