@@ -6,7 +6,7 @@ import 'package:flutter_spinkit/flutter_spinkit.dart';
 import 'package:google_fonts/google_fonts.dart';
 
 class EducationalWidget extends StatefulWidget {
-  const EducationalWidget({Key key}) : super(key: key);
+  const EducationalWidget({Key? key}) : super(key: key);
 
   @override
   _EducationalWidgetState createState() => _EducationalWidgetState();
@@ -16,23 +16,33 @@ class _EducationalWidgetState extends State<EducationalWidget> {
   final scaffoldKey = GlobalKey<ScaffoldState>();
 
   @override
+  void initState() {
+    super.initState();
+
+    WidgetsBinding.instance.addPostFrameCallback((_) => setState(() {}));
+  }
+
+  @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      key: scaffoldKey,
-      backgroundColor: FlutterFlowTheme.of(context).primaryBackground,
-      body: SafeArea(
-        child: GestureDetector(
-          onTap: () => FocusScope.of(context).unfocus(),
-          child: FlutterFlowWebView(
-            url: 'https://covidactnow.org/?s=34452244\n',
-            bypass: false,
-            width: MediaQuery.of(context).size.width,
-            height: MediaQuery.of(context).size.height * 1,
-            verticalScroll: false,
-            horizontalScroll: false,
+    return Title(
+        title: 'educational',
+        color: FlutterFlowTheme.of(context).primaryColor,
+        child: Scaffold(
+          key: scaffoldKey,
+          backgroundColor: FlutterFlowTheme.of(context).primaryBackground,
+          body: SafeArea(
+            child: GestureDetector(
+              onTap: () => FocusScope.of(context).unfocus(),
+              child: FlutterFlowWebView(
+                url: 'https://covidactnow.org/?s=34452244\n',
+                bypass: false,
+                width: MediaQuery.of(context).size.width,
+                height: MediaQuery.of(context).size.height * 1,
+                verticalScroll: false,
+                horizontalScroll: false,
+              ),
+            ),
           ),
-        ),
-      ),
-    );
+        ));
   }
 }
